Move SVGComponents into /assets instead of /svg
More files, less folders.
